refactor(texts): interface
