+++ src/modules/texts/TextEdit/Context/TextEdit.context.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useContext, useState } from "react";
 import { IContextProviderTextEdit, IContextTextEdit } from "./types";
 import { INITIAL_STATE_DATA_TEXT } from "../components/Markdown/constant";
-import { IDataText } from "../components/Markdown/types";
+import { IDataText } from "modules/texts/TextEdit/types";
 
 export const ContextTextEdit = createContext<IContextTextEdit | undefined>(
 	undefined
