+++ src/modules/texts/TextEdit/types/TextEdit.types.ts
@@ -1,0 +1,7 @@
+export interface IDataText {
+	text: string;
+	title_board: string;
+	title_text: string;
+	created_at: string;
+	public: boolean;
+}
