+++ src/modules/texts/TextEdit/types/index.ts
@@ -1,0 +1,2 @@
+export * as default from "./TextEdit.types";
+export * from "./TextEdit.types";
